requirements.txt
certifi==2021.5.30
chardet==4.0.0
idna==2.10
numpy==1.21.0
Pillow==8.2.0
pyTelegramBotAPI==3.8.1
requests==2.25.1
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.8.1+cpu
torchvision==0.9.1+cpu
typing-extensions==3.10.0.0
urllib3==1.26.6